Update Simple Calculator Example and minor observability refactor to speed up launch time
- Updated Dockerfile and README for the Simple Calculator example to reflect correct paths and configurations.
- Moved FileMode class out of FileMixin to speed up bring up time
- Lazy import of aiofile to speed up bring up time
